Methods for S3 classes hashmap and hashset

This page provides an overview of the available methods for hashmap and hashset objects (and for their common parent class hashtable). We list methods based on the general type of task addressed.

Basic read/write operations

insert()

delete()

query()

subsetting_hashtables: `[[`, `[[<-`, `[` and `[<-`

Size of hash table

length.r2r_hashtable()

Other key or value access operations

keys()

values()

has_key, %has_key%

Get/set hashtable properties

hash_fn()

compare_fn()

on_missing_key()

default()

Author

Valerio Gherardi